individuals done now to pipeline
The detection and segmentation interfaces are done.
Individual models (detectron, yolo) are adapted to these as well.
Now, we need to create the engine builder:
- input: backend, task
- output: something that inherits from engine interface

The engine interface is:
class BaseEngine:

    def run(video: Path) -> Iterator:

then for each task we create an engine like:

class DetectionEngine:
    def __init__(self, backend):
        self.runner = choose(backend)

    def run(self, video): self.runner.detect(video)

Then we need to write the visualizer:

class Visual:

    def draw(self, engine_result) -> Frame:

then for each task, we create a visualizer like:

class DetectionVisual(Visual):
    def draw(self, Detection) -> Frame:

Then we combine everything in task runner.

class Pipeline:
    def run(self):

where

class DetectionPipeline(Pipeline):
    def __init__(self, backend):
        self.e = DetectionEngine(backend)
        self.v = DetectionVisual(backend)
    def run(self, video):
        engine_result = self.e.run(video)
        visual = self.v.draw(engine_result)
        save(visual)

Then we do the same in cpp if possible